Avoid deprecated ContentHandler::makeParserOptions()
Having a different ParserOptions for each content model isn't feasible
in an MCR world. And the only thing using this was Wikibase, which has
been fixed to do what it needs in a different way.

This detects a constant defined in I01373b29 to work both before and
after that change.

Bug: T194263
Change-Id: I2260fe7322cace9003c3c17be95cfba8df4bdbbe