Modernize ApiParsoidBatch using ApiResult to generate prettier output (take 2)
This code now matches the output of ApiParse and ApiExpandTemplates in
core when json formatversion 2 is used.  However, Parsoid had
previously been asking for formatversion 2 but expecting formatversion
1 output (sigh).  A patch to Parsoid was necessary to update Parsoid to
handle proper formatversion 2 output from both ApiParsoidBatch and the
core ApiParse and ApiExpandTemplate APIs.

This is a replay of 9e4aea5714e95b7d2cfa75909d71354ca3820130, which
had to be reverted due to T210550.  In this version the problematic
ApiResult::NO_VALIDATE option has been removed, which caused empty
JSON responses when the Kartographer extension attempted to return
binary data as a page property (see T210548).  These properties
were likely originally added in a vain attempt to make JSON
formatversion=1 work; they should be unnecessary now that we
are using formatversion=2.  Further discussion in
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/ParsoidBatchAPI/+/468335/1/includes/ApiParsoidBatch.php#171

Bug: T210550
Depends-On: I8ab01202f700e3f0b87f87ddfd24ce710e668c6d
Change-Id: I0ac404b364b074f11fb752815e53d61d10bd59cf